Proyecto_Final/sprint2/src/requirements.txt
Flask
pycryptodome
aiohttp
requests
